feat: :sparkles: checks if an obj is an instace of that class
This method check if an object is a direct instance of the class of the given.

it should be understood that there is a difference between  type() and isinstnace() method as both
of them tends to be used for similare checks.

if you want to check if an object is strictly an instance of a particular class, the type() function
is the best thing to use. isinstance is used to check if the object is class of the given object or
a superclass of the given object.